requirements.txt
asgiref==3.7.2
Django==4.2.7
django-environ==0.11.2
django-filter==23.3
django-fsm==2.8.1
django-model-utils==4.3.1
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
djangorestframework-recursive==0.1.2
drf-nested-routers==0.93.4
mysqlclient==2.2.0
phonenumbers==8.13.21
Pillow==10.0.1
PyJWT==1.7.1
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
unicef-djangolib==0.7
unicef-restlib==0.7
